requirements.txt
appdirs==1.4.3
attrs==19.2.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
idna==2.8
imageio==2.6.1
lmdb==0.97
numpy==1.17.2
Pillow==6.2.0
requests==2.22.0
toml==0.10.0
urllib3==1.25.6
